frontend/requirements.txt
streamlit==1.44.1
pyyaml==6.0.2
requests==2.32.3
streamlit-autorefresh==1.0.1 